+++ aeryn_core/template_inheritance/__init__.py
@@ -1,0 +1,2 @@
+from .base import TemplateBase, template_base
+__all__ = ['TemplateBase', 'template_base']